big april update

+++ src/main/kotlin/coraythan/keyswap/decks/models/DeckSearchResult.kt
@@ -46,7 +46,7 @@ data class DeckSearchResult(
         val sasRating: Int = 0,
         val synergyRating: Int = 0,
         val antisynergyRating: Int = 0,
-        val metaScores: Map<String, Int> = mapOf(),
+        val metaScores: Map<String, Double> = mapOf(),
         val efficiencyBonus: Double = 0.0,
 
         val totalPower: Int = 0,
